Refactor remove line separator to method in the rich-text package.
